bug fixes and error handling

+++ cogs/scheduled_events/scheduled_events.py
@@ -1,11 +1,10 @@
 import disnake
 from disnake import ApplicationCommandInteraction as AppCmdInter
-from disnake.ext import commands, tasks
+from disnake.ext import commands
 import re
 import yaml
-import asyncio
 from main import SATXBot
-from typing import Dict, List, Tuple
+from typing import Dict
 from .rsvp_view import RsvpView
 
 
@@ -39,7 +38,6 @@ class EventRecords:
         self.event_to_role.pop(event_id)
         await self.rewrite_to_yaml()
 
-
 class ScheduledEventCog(commands.Cog):
     def __init__(self, bot: SATXBot):
         self.bot = bot
@@ -60,7 +58,8 @@ class ScheduledEventCog(commands.Cog):
         await self.announce_event_and_create_thread(event)
 
     @commands.Cog.listener()
-    async def on_guild_scheduled_event_update(self, event_before: disnake.GuildScheduledEvent, event_after: disnake.GuildScheduledEvent):
+    async def on_guild_scheduled_event_update(self, event_before: disnake.GuildScheduledEvent,
+                                              event_after: disnake.GuildScheduledEvent):
         await self.retry_thread_creation(event_before, event_after)
         await self.rename_event_role_and_thread(event_before, event_after)
 
@@ -70,7 +69,8 @@ class ScheduledEventCog(commands.Cog):
         await self.send_late_rsvp(event, subscriber)
 
     @commands.Cog.listener()
-    async def on_guild_scheduled_event_unsubscribe(self, event: disnake.GuildScheduledEvent, subscriber: disnake.Member):
+    async def on_guild_scheduled_event_unsubscribe(self, event: disnake.GuildScheduledEvent,
+                                                   subscriber: disnake.Member):
         await self.unsubscribe_event_role(event, subscriber)
 
     @commands.Cog.listener()
@@ -79,6 +79,40 @@ class ScheduledEventCog(commands.Cog):
         await self.delete_all_rsvp_messages(event)
 
     """ Announcement and Thread Creation """
+
+    @commands.message_command(name="Start Event Management")
+    async def event_management(self, inter: AppCmdInter, event_msg: disnake.Message):
+        await inter.response.defer(ephemeral=True)
+        if inter.author.id not in [self.bot.keys.BOT_OWNER_ID, self.bot.keys.TEST_SERVER_MOD_ID]:
+            await inter.followup.send("Only the bot owner can use this command.", ephemeral=True)
+            return
+
+        event_id = await find_event_id(event_msg.content)
+        if not event_id:
+            await inter.followup.send("No valid event link was found in this message.", ephemeral=True)
+            return
+        event = await inter.guild.fetch_scheduled_event(event_id)
+
+        event_thread_id = event_msg.thread.id
+        event_thread = await inter.guild.fetch_channel(event_thread_id)
+        if not event_thread:
+            await inter.followup.send("There is not a valid event thread connected to this message.",
+                                      ephemeral=True)
+            return
+
+        metro_role_id = await find_metro_role_id(event_msg.content)
+        if not metro_role_id:
+            await inter.followup.send("No metro events role was found in this message.", ephemeral=True)
+            return
+
+        if self.event_records.event_to_thread.get(event_id):
+            await inter.followup.send("This event already is being managed by the bot.", ephemeral=True)
+            return
+
+        event_role = await create_event_role(event.name, event_msg.guild, metro_role_id)
+        event_role = await add_event_role_to_users(event_role, event, metro_role_id)
+        await self.send_help_message(event_msg, event, event_thread, event_role)
+        await inter.followup.send("Event management successful.", ephemeral=True)
 
     async def get_event_message(self, event: disnake.GuildScheduledEvent):
         """
@@ -107,21 +141,16 @@ class ScheduledEventCog(commands.Cog):
 
         # Start event role management
         metro_role_id = self.metroplex_roles.get(event_metroplex)
-        await self.create_event_role_and_help_message(announce_msg, event, event_thread, metro_role_id)
-
-    async def create_event_role_and_help_message(self,
-                                                 announce_msg: disnake.Message,
-                                                 event: disnake.GuildScheduledEvent,
-                                                 event_thread: disnake.Thread,
-                                                 metro_role_id: int):
-        """
-        Create event role and add the role to all event subscribers. Then send the bot help message and add to record.
-        """
         event_role = await create_event_role(event.name, event.guild, metro_role_id)
-        event_subscribers = await event.fetch_users()
-        for subscriber in event_subscribers:
-            await subscriber.add_roles(event_role)
+        event_role = await add_event_role_to_users(event_role, event, metro_role_id)
+        await self.send_help_message(announce_msg, event, event_thread, event_role)
 
+    async def send_help_message(self,
+                                announce_msg: disnake.Message,
+                                event: disnake.GuildScheduledEvent,
+                                event_thread: disnake.Thread,
+                                event_role: disnake.Role
+                                ):
         await event_thread.send(f"<@{event.creator_id}> is the host of this event!")
         await event_thread.send(f"You can ping <@&{event_role.id}> to talk to all interested users.")
         await event_thread.send(f"The host can use the `/send_rsvp` command for an RSVP message to be DMed "
@@ -142,6 +171,7 @@ class ScheduledEventCog(commands.Cog):
             await self.announce_event_and_create_thread(event_after)
 
     """ Event Role """
+
     async def rename_event_role_and_thread(self, event_before: disnake.GuildScheduledEvent,
                                            event_after: disnake.GuildScheduledEvent):
         if event_after.status == disnake.GuildScheduledEventStatus.completed:
@@ -151,32 +181,35 @@ class ScheduledEventCog(commands.Cog):
         if event_before.name == event_after.name:
             return
 
-        if event_after.creator_id != self.bot.keys.BOT_ID:
-            return
-
         event_role = await self.fetch_event_role(event_after.guild_id, event_after.id)
         event_guild = await self.bot.fetch_guild(event_after.guild_id)
-        event_thread = await event_guild.fetch_channel(self.event_records.event_to_thread[event_after.id])
-        event_message = self.bot.get_message(self.event_records.event_to_message[event_after.id])
+        event_thread = await event_guild.fetch_channel(self.event_records.event_to_thread.get(event_after.id))
+
         await event_role.edit(name=event_after.name)
         await event_thread.edit(name=event_after.name)
-        await event_message.edit(content=(await self.get_event_message(event_after)))
+
+        event_message = await self.irl_events_channel.fetch_message(
+            self.event_records.event_to_message.get(event_after.id))
+
+        if event_message.author.id == self.bot.keys.BOT_ID:
+            await event_message.edit(content=(await self.get_event_message(event_after)))
 
     async def delete_event_role(self, event):
-        try:
-            event_guild = await self.bot.fetch_guild(event.guild_id)
-            event_role_id = self.event_records.event_to_role[event.id]
-            event_role = event_guild.get_role(event_role_id)
-            await event_role.delete()
+        event_guild = await self.bot.fetch_guild(event.guild_id)
+        event_role_id = self.event_records.event_to_role.get(event.id)
+        if not event_role_id:
+            return
+        event_role = event_guild.get_role(event_role_id)
+        await event_role.delete()
 
-            await self.event_records.remove_event(event.id)
-        except KeyError:
-            pass
+        await self.event_records.remove_event(event.id)
 
     async def add_role_and_ping_in_thread(self, event: disnake.GuildScheduledEvent, subscriber: disnake.Member):
         if subscriber.id == event.creator_id:
             return
-        event_thread = self.bot.get_channel(self.event_records.event_to_thread[event.id])
+        event_thread = self.bot.get_channel(self.event_records.event_to_thread.get(event.id))
+        if not event_thread:
+            return
         await event_thread.send(f"<@{subscriber.id}> is interested in **{event.name}**!")
         event_role = await self.fetch_event_role(event.guild_id, event.id)
         await subscriber.add_roles(event_role)
@@ -186,8 +219,6 @@ class ScheduledEventCog(commands.Cog):
         await subscriber.remove_roles(event_role)
 
     async def fetch_event_role(self, guild_id: int, event_id: int) -> disnake.Role:
-        while not self.event_records.event_to_role.get(event_id):
-            await asyncio.sleep(0.1)
         event_role_id = self.event_records.event_to_role.get(event_id)
         guild = await self.bot.fetch_guild(guild_id)
         event_role = guild.get_role(event_role_id)
@@ -197,7 +228,9 @@ class ScheduledEventCog(commands.Cog):
 
     async def send_rsvp_message(self, event: disnake.GuildScheduledEvent,
                                 subscriber: disnake.Member, rsvp_list_message: disnake.Message) -> disnake.Message:
-        event_thread = self.bot.get_channel(self.event_records.event_to_thread[event.id])
+        event_thread = self.bot.get_channel(self.event_records.event_to_thread.get(event.id))
+        if not event_thread:
+            raise LookupError(f"Event thread {event.id} not found.")
         event_creator = await self.bot.fetch_user(event.creator_id)
 
         rsvp_view = RsvpView(event, event_thread, rsvp_list_message, subscriber, event_creator)
@@ -237,10 +270,14 @@ class ScheduledEventCog(commands.Cog):
         if self.rsvp_messages.get(event.id) is None:
             # RSVP messages have not been sent out yet.
             return
-        event_thread = self.bot.get_channel(self.event_records.event_to_thread[event.id])
+        event_thread = self.bot.get_channel(self.event_records.event_to_thread.get(event.id))
+        if not event_thread:
+            return
         event_creator = await self.bot.fetch_user(event.creator_id)
 
-        rsvp_list_message_id = self.rsvp_list_messages[event.id]
+        rsvp_list_message_id = self.rsvp_list_messages.get(event.id)
+        if not rsvp_list_message_id:
+            return
         rsvp_list_message = self.bot.get_message(rsvp_list_message_id)
 
         rsvp_view = RsvpView(event, event_thread, rsvp_list_message, subscriber, event_creator)
@@ -252,13 +289,14 @@ class ScheduledEventCog(commands.Cog):
             # RSVP messages were not sent
             return
         # Delete each individual list
-        for rsvp_msg_id in self.rsvp_messages[event.id]:
+        for rsvp_msg_id in self.rsvp_messages.get(event.id):
             msg = self.bot.get_message(rsvp_msg_id)
             await msg.delete()
         self.rsvp_messages.pop(event.id)
         self.rsvp_list_messages.pop(event.id)
 
     """ Bot Initialization """
+
     async def read_from_event_records(self):
         """
         Reads all the prior event records from the event_records.yaml file and adds them to the bot's event records
@@ -281,42 +319,14 @@ class ScheduledEventCog(commands.Cog):
         guild = await self.bot.fetch_guild(self.bot.keys.TEST_SERVER_ID)
         events = await guild.fetch_scheduled_events()
         for event in events:
-            role_id = self.event_records.event_to_role[event.id]
+            role_id = self.event_records.event_to_role.get(event.id)
+            if not role_id:
+                continue
             event_role = guild.get_role(role_id)
             users = await event.fetch_users()
             for user in users:
                 if not user.get_role(role_id):
                     await self.add_role_and_ping_in_thread(event, user)
-
-    @commands.message_command(name="Start Event Management")
-    async def start_event_management(self, inter: AppCmdInter, event_msg: disnake.Message):
-        if inter.author.id not in [self.bot.keys.BOT_OWNER_ID, self.bot.keys.TEST_SERVER_MOD_ID]:
-            await inter.response.send_message("Only the bot owner can use this command.", ephemeral=True)
-            return
-
-        event_id = await find_event_id(event_msg.content)
-        if not event_id:
-            await inter.response.send_message("No valid event link was found in this message.", ephemeral=True)
-            return
-        event = await inter.guild.fetch_scheduled_event(event_id)
-
-        event_thread = event_msg.thread
-        if not event_thread:
-            await inter.response.send_message("There is not a valid event thread connected to this message.",
-                                              ephemeral=True)
-            return
-
-        metro_role_id = await find_metro_role_id(event_msg.content)
-        if not metro_role_id:
-            await inter.response.send_message("No metro events role was found in this message.", ephemeral=True)
-            return
-
-        if self.event_records.event_to_thread.get(event_id):
-            await inter.response.send_message("This event already is being managed by the bot.", ephemeral=True)
-            return
-
-        await self.create_event_role_and_help_message(event_msg, event, event_thread, metro_role_id)
-        await inter.response.send_message("Event management successful.", ephemeral=True)
 
 
 async def get_event_link(guild_id: int, event_id: int) -> str:
@@ -339,7 +349,23 @@ async def find_metro_role_id(input_string: str) -> int:
 
 async def create_event_role(event_name: str, guild: disnake.Guild, metro_role_id: int) -> disnake.Role:
     metro_role = guild.get_role(metro_role_id)
+    roles = await guild.fetch_roles()
+    if event_name in roles:
+        raise Exception("Event likely already has a role")
+
     return await guild.create_role(name=event_name, color=metro_role.color, mentionable=True)
+
+
+async def add_event_role_to_users(event_role: disnake.Role,
+                                  event: disnake.GuildScheduledEvent,
+                                  metro_role_id: int):
+    """
+    Create event role and add the role to all event subscribers. Then send the bot help message and add to record.
+    """
+    event_subscribers = await event.fetch_users()
+    for subscriber in event_subscribers:
+        await subscriber.add_roles(event_role)
+    return event_role
 
 
 async def get_empty_rsvp_embed(event_creator_id, event_name):
